Introducing sub strip resolution (center-of-charge)

+++ tools/include/Hit.h
@@ -7,7 +7,7 @@
 class Hit
 {
     private:
-        int _center;
+        double _center;
         int _left;
         int _right;
         double _sig;
@@ -15,13 +15,13 @@ class Hit
         
         void cpy(const Hit &h);
     public:
-        Hit(int c=0, int l=0, int r=0, double s=0, double s_c=0);
+        Hit(double c=0, int l=0, int r=0, double s=0, double s_c=0);
         Hit(const Hit &h);
         ~Hit();
         
         Hit &operator=(const Hit &h);
         
-        int center() const { return _center; }
+        double center() const { return _center; }
         int left() const { return _left; }
         int right() const { return _right; }
         double signal() const { return _sig; }
